move object property heuristics to heuristics.rs
